HOTFIX: Support advanced union features in flatbuffers 1.10.0
This allows us to use flatc --binary in flatbuffers 1.10.0, which
we are stuck on until we ditch gRPC+flatbuffers.

We have a schema that has a union with a struct type inside of it,
which is supported by the binary schema format but not flagged
appropriately in this check.

This works fine in flatbuffer 2.0.0 which is what we will
switch to later.