Fix toolchain collision
Hunter computes a toolchain-id using the compiler vendor, version, and
compile flags. The idea is that if you change some flag that will
change the ABI, it will compute a new toolchain-id which will rebuild
dependencies.

This toolchain-id is computed by taking all of those inputs and
computing a SHA1 digest, and then truncating it to the first 7
characeters.

On Debian 12 I'm finding that this constantly collides when switching
branches, which requires me to delete the build directory every time I
change to a new branch.

This change truncates the CONFIG-ID, TOOLCHAIN-ID and HUNTER-ID to 9
characters instead of 7, which resolves the problem for me.